Helper to include user/flash on every template
I got tired of repeating myself, including friend/identity and (:flash req) on every template. It was especially annoying when I had to include the identity in the base template *and* the snippet. The `wrap-view` helper takes the request object and a snippet call, and assocs the identity and flash values with the parameters to that snippet and the base template.