Custom view transformations and helpers
I wrote a few functions to help transform enlive templates based on patterns that were showing up in my code. The most common pattern is showing or hiding a selected node, depending on whether a user is logged in. Another less common pattern is modifying link attributes, depending on whether a submitted post points to an external URL.