Rename notes back to comments.
I had originally renamed the "comment" table to "note", to avoid collision with the built-in comment function. After working with it, it makes more sense to use "comments" (the word most commonly used for that sort of text reply). Besides, it'll overwrite the comment function already in the namespace, and since I never see myself using that function, it doesn't really matter.